@@ -4,4 +4,6 @@
  * For now, just export the client, which will work just fine. But, later, it'll
  * be important to give out nodes evenly from across the DHT.
  */
-module.exports = require('./client')
+
+import client from './client.js'
+export default client
