@@ -1,6 +1,6 @@
-const common = require('./common')
-const DHT = require('../')
-const test = require('tape')
+import common from './common.js'
+import DHT from '../index.js'
+import test from 'tape'
 
 test('explicitly set nodeId', t => {
   const nodeId = common.randomId()
